pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "oxylus-erp"
description = "Oxylus applications for enterprise and organisations management system."
version = "0.0.1"
authors = ["Thomas Baquet <thomas oxylus app>"]
readme = "README.rst"
license = "GPL-3.0-or-later"
packages = [
    { include = "ox_erp" },
]
include = [
    # assets are provided as npm packages
    # {path="assets", format="sdist"},
    {path="docs", format="sdist"},
    {path="tests", format="sdist"},
]
exclude = [
]

classifiers = [
    "Framework :: Django",
    "Framework :: Oxylus",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
]


[tool.poetry.dependencies]
python = "^3.13"
Django = "^5"
djangorestframework = "^3"
django-filter = "*"
# django-cities = "^0.6.2"
ipython = "*"
pycountry = "^24.6.1"
phonenumbers = "^9.0.4"
python-stdnum = "^2.0"

oxylus = { path = "../oxylus" }
django-caps = "*"

[tool.poetry.group.dev.dependencies]
pre-commit = "*"
black = "*"
ruff = "*"
pytest = "*"
pytest-django = "*"
pytest-asyncio = "^1.0.0"
pytest-mock = "^3.14.1"

[tool.poetry.group.docs.dependencies]
Sphinx = "*"
sphinx-js = "*"
sphinx-rtd-theme = "*"
sphinx-autodoc-typehints = "*"


# [tool.pytest.ini_options]
# DJANGO_SETTINGS_MODULE = "ox.settings.test"
# python_files = ["tests.py", "test_*.py", "*_tests.py"]


[tool.black]
line-length = 120
exclude = '''
    /(
          \.egg
        | \.git
        | \.hg
        | \.tox
        | \._build
        | \.build
        | \.bulk-out
        | \.dist
        | \.__pycache__
        | \.venv
        | \.migrations
        | \.static
        | \.ox/settings
        | \.scripts
    )
'''


[tool.ruff]
line-length = 120
exclude = [
    "egg",
    "git",
    "hg",
    "tox",
    "_build",
    "build",
    "dist",
    "__pycache__",
    "venv",
    "*/migrations",
    "*/static",
    "ox/settings",
    "scripts/",
]


[tool.semantic_release]
version_variables = ["ox/__init__.py:__version__"]
version_toml = [
    "pyproject.toml:tool.poetry.version",
]
